feat: hooks.server.ts 및 hooks.ts 추가
hooks.server.ts에서 paraglideMiddleware를 사용하여 서버 측 오류 처리를 추가했음. hooks.ts에서 deLocalizeUrl을 사용하여 URL을 변환하는 reroute 기능을 추가했음. boiler_plates의 hooks 파일들을 base 라이브러리의 hooks 파일로 리팩토링했음. package.json에 hooks 관련 경로를 추가했음.
