+++ apps/boiler_plates/src/hooks.client.ts
@@ -1,0 +1,1 @@
+export * from '@library/base/hooks.client.ts'
