+++ apps/boiler_plates/src/hooks.server.ts
@@ -1,18 +1,1 @@
-import { paraglideMiddleware } from '@library/paraglide/server.js'
-import type { Handle } from '@sveltejs/kit'
-
-// creating a handle to use the paraglide middleware
-const paraglideHandle: Handle = ({ event, resolve }) =>
-	paraglideMiddleware(
-		event.request,
-		({ locale }) => {
-			return resolve(event, {
-				transformPageChunk: ({ html }) => html.replace('%lang%', locale),
-			})
-		},
-		{
-			disableAsyncLocalStorage: true,
-		},
-	)
-
-export const handle: Handle = paraglideHandle
+export * from '@library/base/hooks.server.ts'
