+++ apps/boiler_plates/src/hooks.ts
@@ -1,6 +1,1 @@
-import { deLocalizeUrl } from '@library/paraglide/runtime.js'
-import type { Reroute } from '@sveltejs/kit'
-
-export const reroute: Reroute = (request) => {
-	return deLocalizeUrl(request.url).pathname
-}
+export * from '@library/base/hooks.ts'
