+++ libraries/base/src/hooks.ts
@@ -1,0 +1,6 @@
+import { deLocalizeUrl } from '@library/paraglide/runtime.js'
+import type { Reroute } from '@sveltejs/kit'
+
+export const reroute: Reroute = (request) => {
+	return deLocalizeUrl(request.url).pathname
+}
